+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Fixed creating a buffer view not checking the `min_texel_buffer_offset_alignment` limit.
 - Added support for loading the `VK_EXT_debug_marker` extension and adding debug markers to
   `UnsafeCommandBufferBuilder`
+- Changed `GraphicsPipelineBuilder` to default to a buffer-less vertex input.
+- Deprecated `pipeline::input_assembly::InputAssembly`.
 
 # Version 0.7.1 (2017-09-28)